requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
albumentations==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bash_kernel==0.9.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
cachetools==5.3.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
efficientnet==1.0.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
flatbuffers==23.3.3
fonttools==4.40.0
fqdn==1.5.1
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1
grpcio==1.51.3
h5py==3.8.0
idna==2.8
image-classifiers==1.0.0
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-archive==3.3.4
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-http-over-ws==0.0.8
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
kaggle==1.5.15
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nbzip==0.1.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.23.5
nvidia-ml-py==11.525.131
oauthlib==3.2.2
opencv-contrib-python-headless==4.8.0.74
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patchify==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.15.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfunctional==1.4.3
Pygments==2.15.1
PyGObject==3.36.0
pyparsing==3.1.0
pyrsistent==0.19.3
pytest==7.3.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
segmentation-models==1.0.1
Send2Trash==1.8.0
shapely==2.0.1
six==1.14.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
torch==1.8.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.25.8
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
